refactor: DisplaySettingsScreen→Page リネームと eew/ui のディレクトリ整理
- DisplaySettingsScreen クラス名を DisplaySettingsPage に変更（flutter-rules.mdc 命名規則統一）
- eew_details_by_event_id_page.dart を ui/screen/ から ui/page/ に移動（ファイル名は _page.dart だがディレクトリが不整合だったため修正）
- router.dart の対応する import・参照をあわせて更新
