+++ app/lib/core/router/router.dart
@@ -12,7 +12,7 @@ import 'package:eqmonitor/feature/earthquake_history/ui/earthquake_history_page.
 import 'package:eqmonitor/feature/earthquake_replay/ui/earthquake_replay_page.dart';
 import 'package:eqmonitor/feature/earthquake_search/data/model/earthquake_search_parameter.dart';
 import 'package:eqmonitor/feature/earthquake_search/ui/earthquake_search_result_page.dart';
-import 'package:eqmonitor/feature/eew/ui/screen/eew_details_by_event_id_page.dart';
+import 'package:eqmonitor/feature/eew/ui/page/eew_details_by_event_id_page.dart';
 import 'package:eqmonitor/feature/home/ui/page/home_map_layer_page.dart';
 import 'package:eqmonitor/feature/kyoshin_monitor/page/kyoshin_monitor_about_observation_network_page.dart';
 import 'package:eqmonitor/feature/kyoshin_monitor/page/kyoshin_monitor_about_page.dart';
@@ -345,7 +345,7 @@ class DisplayRoute extends GoRouteData with $DisplayRoute {
 
   @override
   Widget build(BuildContext context, GoRouterState state) =>
-      const DisplaySettingsScreen();
+      const DisplaySettingsPage();
 }
 
 class NotificationSettingsRoute extends GoRouteData
