+++ app/lib/feature/eew/ui/page/eew_details_by_event_id_page.dart
